116-118

+++ backend/routes/posts.js
@@ -27,31 +27,36 @@ const storage = multer.diskStorage({
   }
 })
 
-router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
+router.post(
+  "",
+  checkAuth,
+  multer({storage: storage}).single("image"),
+  (req, res, next) => {
   const url = req.protocol + '://' + req.get('host');
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId,
   });
   post.save().then(createdPost => {
     res.status(201).json({
       message: "Post added successfully",
       post: {
+        ...createdPost,
         id: createdPost._id,
-        title: createdPost.title,
-        content: createdPost.content,
-        imagePath: createdPost.imagePath,
       }
     });
   });
 });
 
-router.put('/:id', checkAuth, multer({storage: storage}).single("image"),
+router.put(
+  "/:id", checkAuth,
+  multer({storage: storage}).single("image"),
   (req, res, next) => {
     let imagePath = req.body.imagePath;
     if (req.file) {
-      const url = req.protocol + '://' + req.get('host');
+      const url = req.protocol + "://" + req.get("host");
       imagePath = url + "/images/" + req.file.filename;
     }
   const post = new Post({
@@ -61,9 +66,14 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"),
     imagePath: imagePath
   });
     console.log("Post updated Info: ", post);
-    Post.updateOne({_id: req.params.id}, post).then(result => {
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post)
+      .then(result => {
     console.log(result);
-    res.status(200).json({message: "Post updated successfully!"});
+    if(result.modifiedCount > 0) {
+      res.status(200).json({message: "Post updated successfully!"});
+    } else {
+      res.status(401).json({message: "Not authorized from post rounds method put"});
+    }
   });
 });
 
@@ -75,22 +85,20 @@ router.get("", async (req, res, next) => {
     const postQuery = Post.find();
 
     if (pageSize && currentPage) {
-      postQuery
-        .skip(pageSize * (currentPage - 1))
-        .limit(pageSize);
+      postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
     }
 
-    const documents = await postQuery;
+    const fetchedPost = await postQuery;
     const count = await Post.countDocuments();
 
     res.status(200).json({
       message: "Posts fetched successfully!",
-      posts: documents,
+      posts: fetchedPost,
       maxPosts: count
     });
   } catch (error) {
     res.status(500).json({
-      message: "Fetching posts failed!",
+      message: "Fetching posts failed from post routs, line 94!",
       error: error
     });
   }
@@ -107,9 +115,13 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json({message: "Post deleted!"});
+    if(result.deletedCount > 0) {
+      res.status(200).json({message: "Post deleted!"});
+    } else {
+      res.status(401).json({message: "Not authorized from post routes method delete"});
+    }
   });
 });
